adsExApp/Db/adsTestAsyn.db: out records with READBACK
There are 2 ways to use the READBACK:
Either by using "I/O Intr" or by a polling with 1 Hz.
When an output record is processed, and the value pushed
into the PLC, the PLC may decide to ignore the change in the next PLC
cycle. And set back the variable back to where it was.
As there is "no change" from a PLC perspective, the EPICS record
and the PLC variable get of synch.
There is no notifaction send via ADS (2 would have been better).
As a workaround, the 1Hz poller can be used. Document this better.